Fix two prop-types errors (#825)
In both cases, the problem is that due to recent changes in how
translations are handled, the translation passed to the
render-function content of `<FormattedMessage>` can sometimes be an
array of a single string, rather than just a string. We allow for this
case by checking the type of passed translation.