feat: AL sorting
Sort agreement lines by name then reference then UUID in the UI. Switches from manifest object to function returning an object where the sort is an array. (The fact this is necessary makes me sad... roll on react-query). This allows us to bake 3 separate 'sort's into the agreement line query. To go with this we have to switch from "filters: "owner=:{id}"" to the more explicit "filters: `owner=${params?.id}`" using the params handed to us by the manifest

ERM-1931