fix: Agreement Internal Contact fetch
Ensure initial fetch for users, so that subsequent agreement fetch will then refresh that call (stripes-connect not returning original params would somehow mean that an agrement refresh down the road would not then fetch that resource. This fixes that)
This work includes a tweak to how the `fetch` param is calculated, to avoid an unnecessary initial call to users before the agreement has loaded. This way the resource exists in stripes-connect ready for refreshing, but the call itself is not made.

refs ERM-2063