fix: useKiwtFieldArray onAddField used incorrectly
We previously had `onClick={onAddField}` instead of `onClick={() => onAddField()}`. The first one is a javascript shortcut for: `onClick={(e) => onAddField(e)}`, and for this function if you pass a prop in it will use that as the starting point for the field, so it was attempting to add a click event to the field, instead of an empty linked license object.

Also made babel core dep ^ instead of static

ERM-2299