Release 9.0.4 (#1270)
* fix: Prevent items feedback loop

Added in special handling for items which removes all information from `owner` property except id. Is not an ideal fix, but future versions have performance changes in place already, so this is the fastest viable way to get this up and running

ERM-3076

* fix: fixed cases where there is no owner

Some items have no owner, in which case trying to manipulate the owner object would fail. Fixed to return `owner: null` in that case.

ERM-3076

* chore: Release 9.0.4

Added CHANGELOG section and bumped version to 9.0.4

refs ERM-3076, ERM-3077

* build: Github actions node 18

Changed github actions to build with node engine 18 instead of 16 to allow for release with newer patch versions of packages

* build: Revert node18 and tweak deps

Reverted node 18 change, since Orchid was not set up for that, and instead locked buggy version of rehooks/local-storage, and locked version of stripes-testing to Orchid's dep.

ERM-3077