fix: Fixes to accommodate stripes changes
Stripes changed to use `formatter` prop in Selection for rendering display as well as the options. This means that a `null` option is now a legitimate use case, which we had not protected for. Added in null safety here, and this also means that the display on selection is equivalent to the options render, which might not be desired but at least this works.

ERM-3362