ERM-3392 Changes to the field content type are not saved, but a success message is displayed (#1368)
* ERM-3392 Changes to the field content type are not saved, but a success message is displayed

* change agreementContentTypes initialValues on submit

* * revert submit handler changes
* fix field name, add parse and format

* fix button

* chore: Change over to just "id" in select to avoid having to do loads of field level stuff

In general, Claudia's solution is "nicer" in that it prevents ANY incorrect information from appearing in the PUT.

However, it ends up doing a bunch of extra work inside the Select which we would then need to replicate everywhere to get the "best" solution consistently.

We could (and an argument exists should) do this in a RefdataSelect type component, but then that would require special testing etc etc, and to be used in a very specific way. The easier fix (which is already used in some other places) is simply to use the id _only_.

This makes some of the "only display refdata in use" logic a bit more complex, so we maybe ought to improve that and centralise it as well.

ERM-3392

---------

Co-authored-by: EthanFreestone <54310740+EthanFreestone@users.noreply.github.com>
Co-authored-by: Ethan Freestone <ethan.freestone@k-int.co.uk>